# docs/swagger.yaml
basePath: /
definitions:
...
  handler.SugestaoValor:
    properties:
      label:
        type: string
      valor:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API do sistema CronFlow para agendamento e monitoramento de execução
    de tarefas distribuídas em background.
  title: CronFlow API
  version: "1.0"
paths:
...
  /v1/pix/qr:
    get:
      description: Retorna a string do QR Code no formato Data URI Base64 e o payload
        copia-e-cola correspondente para a chave Pix especificada e o valor solicitado.
      parameters:
      - description: 'Valor da doação (ex: 1.50, padrão 1.00)'
        in: query
        name: valor
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: QR Code base64 e payload copia-e-cola
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Não autenticado
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Erro ao computar QR Code ou payload
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Gerar QR Code PIX Dinâmico
      tags:
      - PIX Apoio
  /v1/pix/valores:
    get:
      description: Retorna as opções rápidas e sugeridas de valores para fazer uma
        doação de apoio via PIX.
      produces:
      - application/json
      responses:
        "200":
          description: Opções sugeridas de apoio
          schema:
            items:
              $ref: '#/definitions/handler.SugestaoValor'
            type: array
        "401":
          description: Não autenticado
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Listar valores sugeridos para apoio PIX
      tags:
      - PIX Apoio
securityDefinitions:
  ApiKeyAuth:
    description: Tipo "Bearer <sua-api-key>" para se autenticar.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"